+++ CHANGELOG.md
@@ -27,6 +27,13 @@
 
 ### Added
 
+- [#2741](https://github.com/ChainSafe/forest/issues/2741): Support importing
+  zstd compressed snapshot car files
+- [#2741](https://github.com/ChainSafe/forest/issues/2741): Support fetching
+  zstd compressed snapshots with filecoin provider via `--compressed` option
+- [#1454](https://github.com/ChainSafe/forest/issues/1454): Added state
+  migration support for NV18.
+
 ### Changed
 
 ### Removed